Add Rails concern with common functionality
`GovukPersonalisation::AccountConcern` now holds accounts-related
functionality extracted from `frontend` and `finder-frontend` that these
two apps have in common.

To test that the concern behaves as expected when included in apps that
use this gem, a test app with three routes and a simple
`SessionsController` has been used here.

Trello card: https://trello.com/c/40YbnkRp/783-add-a-rails-concern-to-govukpersonalisation

Co-Authored-By: Alan Gabbianelli <Alan.Gabbianelli@digital.cabinet-office.gov.uk>
Co-Authored-By: Huw Diprose <Huw.Diprose@digital.cabinet-office.gov.uk>