Always set `Vary: GOVUK-Account-Session` response header
Right now we leave it up to individual apps to decide whether they
want to set

    Vary: GOVUK-Account-Session

or

    Cache-Control: no-store

This change means that apps won't need to worry about the `Vary`
header any more.

Setting just the `Vary` header is good when we want the user's browser
to be able to cache the page.  For example, the Transition Checker
results page can be cached, as the querystring contains all the data:
the only dynamic content on the page is the "these results are
different to the ones in your account" box, and it's no big deal if
that's missing due to the user seeing a cached copy.

Setting the `Cache-Control` header is good when we don't want to cache
the page at all, for example the `/sign-in` and `/sign-out` routes,
which manipulate the user's session.

But there's no downside to *always* setting the `Vary` header:

1. If the app only wants to set the `Vary` header, now they don't have
to (because the concern is doing it)

2. If the app wants to set the `Cache-Control` header, it can still do
that, and the `Vary` header becomes a no-op (because `Vary` restricts
when caching can happen, and the app is disabling caching)

There is no case in which we *need* `Cache-Control: no-store` to be
set and for `Vary: GOVUK-Account-Session` to be unset.  But even if
there were, the app could just use `skip_before_action` to achieve
that.