Add a module and methods for managing flash messages
The flash is a special piece of per-request state in the Rails
session, useful for showing success or failure messages, or other
confirmation-style notices.  Since flash messages are in the Rails
session, they cannot be used across microservices (without sharing the
session encryption key). But that is a feature we would like.

To allow sending flash messages for account purposes between
apps (e.g., a user signs up to email notifications for a page, then
gets redirected back to that page and it has a success banner), we
have decided to add them to the account session.

This commit adds a module for managing flash messages in the session,
and adds some helper methods to the `AccountConcern` to make it easy
to use.

See https://github.com/alphagov/account-api/blob/main/docs/api.md