Handle three-part sessions: CSRF token, header, and flash
We're adding a third, optional, component to the session: a CSRF
token.  I've called it an "authenticity token" to match the
terminology Rails uses for its CSRF tokens.

As with the flash, it'll be separated from the session ID by a "$$".
After splitting on "$$", valid session header will be one of:

- 0 components: `[                         ]`
- 1 component:  `[csrf_token               ]`
-               `[            header       ]`
- 2 components: `[csrf_token, header       ]`
-               `[            header, flash]`
- 3 components: `[csrf_token, header, flash]`

To disambiguate a CSRF token from a header, CSRF tokens start with
"csrf_".

The use-case for this feature will be:

1. Authenticating a user: the OAuth state will be saved as the CSRF
token in the session, and checked in the callback.

2. Passing a user from the auth callback to another part of GOV.UK:
the state will be passed as a query param and checked in the app.

(1) is so that an attacker can't just link a user to
www.gov.uk/account/sign-in/callback?code=...&state=... to log them in
to an account (or, trick them into authenticating and sending the
attacker such a URL).

(2) is so that an attacker can't just link a user to
www.gov.uk/make-a-new-email-subscription?topic=... (or whatever) to
create a new email subscription if they're already logged in.
Normally we'd protect such pages by making them POST endpoints and
using the default Rails CSRF protection but, since we need to be able
to handle GETs, we need to implement our own thing.