Add a verify_account_authenticity_token controller method
This should be called in methods which make some state change and are
exposed to GET requests, as these methods will not be protected by
Rails' built-in CSRF protection.

We have such methods because a common pattern with OAuth is:

1. Try to do a thing
2. Fail, because the user's session is missing / invalid
3. Redirect them to authenticate
4. Redirect them back to the original page
5. Do the thing

So we need the page which does the thing to be available for GET
requests, which means we need CSRF protection for it.