Add `GovukPersonalisation::Urls` module
Currently we have ad-hoc approaches to generating account URLs in our
frontend apps.  Some use env vars with the full URL in.  Some
construct paths based on the account-manager's URL.  Some assume URLs.

It's even worse when we want to construct links which work both in
local development (where they need to link to `foo.dev.gov.uk`) and in
production (where they need to link to `www.gov.uk`).

So, let's put all of the URL generation logic in this gem, and re-use
it in different apps.  For now, I've added the URLs which we use in
the account template in govuk_publishing_components.  Then we can
change govuk_publishing_components to depend on this gem, update the
template, and set the URL env vars on Integration to point to Digital
Identity.

We can migrate other URLs into this gem as needed.