Add redirect_with_analytics to ControllerConcern
As we look to hand over portions of Single Sign On to the Digital
Identity team, we increase the number of times a user travels between
gov.uk subdomains.

We store cookie consent within a session which is scoped to a domain.
Before this change, a user who had accepted cookies on one page, would
travel to another page, perhaps still "within their account" and get
bugged with yet another cookie banner. This is a bad user experience.

We already added _ga params to cross domain redirects where the user had
consented. This gives us a more flexible helper that allows to
additionally add their cookie consent choice, so that other domains can
take the appropriate action.

In aid of this we extend the ControllerConcern with a
redirect_with_analytics method, this will take relevant parameters of
the user's existing URL and add them to a relevant redirect.

We can extend this in future if additional parameters are required
without requiring apps to change their code.

Co-authored-by: Alex Whitehead-Smith <alex.smith@digital.cabinet-office.gov.uk>
Co-authored-by: Huw Diprose <huw.diprose@digital.cabinet-office.gov.uk>