Make session-change events uncacheable
We already do this in our Fastly configuration, but that isn't a full
solution, as our own Varnish may cache such a response.

This hasn't been an issue so far, as all session-changing pages have
been explicitly uncacheable (or had `Vary: GOVUK-Account-Session`);
but it is a problem for pages with flash messages, as the controller
concern will strip the flash from the cookie, and so if that response
gets cached, we'll switch subsequent users who visit the same
page (perhaps with the same flash message, if `Vary:
GOVUK-Account-Session-Flash` is set) to that first user's account.