Update to Rails 7
We previously brought `require "sprockets/railtie"` into
`application.rb` in Rails 7 sprokets is [now an optional dependency][1].

If we still need this we should add `sprockets-rails` into the gemfile.
However given sprokets is an asset pipeline, and this gem is mostly a
collection of concerns and helpers. It is irrelivant here.

[1]:https://edgeguides.rubyonrails.org/7_0_release_notes.html#railties-notable-changes