Upgrade ruby to 2.7.5
Updating `.ruby-version` only changes the version of Ruby `rbenv` uses
so consumers of this gem won't be affected at all. Therefore we don't
need to add a changelog entry or bump the version.

We still test on CI with Ruby 2.6 and 2.7 so there's also no need to
update the minimum version in the gemspec.

Co-authored-by: Huw Diprose <huw.diprose@digital.cabinet-office.gov.uk>