Set account home page to Digital Identity
We are in the process of moving the account home page from
`www.gov.uk/account/home` (served by Frontend) to an app on the DI side,
which is part of separating out central account features from the
`www.gov.uk` relying party.

To do so, we'll update `govuk_personalisation` with the new link and
change the controller in Frontend that currently serves that page to
redirect on to the new one. This allows us to do an environment aware
redirect.

Apps currently have the `YOUR_ACCOUNT` environment variables set (see
https://github.com/alphagov/govuk-puppet/pull/11728), which override
the URI generated by this library. This means this change is safe to be
released and it won't affect where apps send users yet.

Once we're ready, we'll make another change in Puppet and set the
environment variables to point to the DI URI.